refactor: Move User-page
